Make hasMatchingLabels generic
