+++ controllers/machine_helpers_test.go
@@ -127,3 +127,74 @@ func Test_getActiveMachinesInCluster(t *testing.T) {
 		})
 	}
 }
+
+func TestMachineHealthCheckHasMatchingLabels(t *testing.T) {
+	testCases := []struct {
+		name     string
+		selector metav1.LabelSelector
+		labels   map[string]string
+		expected bool
+	}{
+		{
+			name: "selector matches labels",
+
+			selector: metav1.LabelSelector{
+				MatchLabels: map[string]string{
+					"foo": "bar",
+				},
+			},
+
+			labels: map[string]string{
+				"foo": "bar",
+			},
+
+			expected: true,
+		},
+		{
+			name: "selector does not match labels",
+
+			selector: metav1.LabelSelector{
+				MatchLabels: map[string]string{
+					"foo": "bar",
+				},
+			},
+
+			labels: map[string]string{
+				"no": "match",
+			},
+			expected: false,
+		},
+		{
+			name:     "selector is empty",
+			selector: metav1.LabelSelector{},
+			labels:   map[string]string{},
+			expected: false,
+		},
+		{
+			name: "seelctor is invalid",
+			selector: metav1.LabelSelector{
+				MatchLabels: map[string]string{
+					"foo": "bar",
+				},
+				MatchExpressions: []metav1.LabelSelectorRequirement{
+					{
+						Operator: "bad-operator",
+					},
+				},
+			},
+			labels: map[string]string{
+				"foo": "bar",
+			},
+			expected: false,
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			g := NewWithT(t)
+
+			got := hasMatchingLabels(tc.selector, tc.labels)
+			g.Expect(got).To(Equal(tc.expected))
+		})
+	}
+}
